Tidied for makefile cleanup

+++ src/basic.c
@@ -114,10 +114,10 @@ bool checkFormulaDividend(__int128_t dividend) {
 bool tryBasic(uint64_t k) {
 	bool result = false;
 	int64_t negBound = -2 * sqrtl(k);
-    uint64_t _6k = 6 * k;
+    int64_t _6k = 6 * k;
     __int128_t dividend = -1;
 	for (int64_t z = negBound; dividend <= 0 && !result; z++) {
-		dividend = (__int128_t)z * z * z - z - (__int128_t)_6k;
+		dividend = (__int128_t)z * z * z - z - _6k;
 		if (checkFormulaDividend(dividend)) {result = true;}
 	}
 	return result;
